backend_app_requirements_Version2.txt
google-adk>=0.1.0
google-cloud-firestore>=2.16.0
google-cloud-secret-manager>=2.20.0
google-cloud-logging>=3.10.0
aiohttp>=3.9.5
pydantic>=2.7.0
uvloop>=0.19.0
protobuf>=4.25.3
google-generativeai>=0.6.0
google-api-python-client>=2.132.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0